Added signal for plot zoom
A signal is emitted when the user zooms in on a plot. Methods for changing plot boundaries have also been added.

+++ GuiFramework/Include/Widgets/PlotSeries1D.h
@@ -32,5 +32,5 @@ public:
 
 	void setBounds(float lower, float upper);
 
-	void setHead(float head);
+	void setHead(int head);
 };
